Db rework

See merge request xkrejci/pv-168-easy-food!16

+++ src/main/java/cz/muni/fi/pv168/project/storage/sql/dao/UnitDao.java
@@ -238,7 +238,7 @@ public class UnitDao implements DataAccessObject<UnitEntity> {
                 resultSet.getString("unitName"),
                 resultSet.getString("abbreviation"),
                 resultSet.getDouble("amount"),
-                resultSet.getLong("baseUnitId")
+                resultSet.getInt("baseUnitId")
         );
     }
